finish codeshare
